refactor(core): rename debug to contextual-ui
