+++ src/core/contextual-ui/EntityOtuline.tsx
