+++ src/core/contextual-ui/EntityOutline.test.tsx
