+++ src/core/contextual-ui/IDebuggerEntity.tsx
