+++ src/core/contextual-ui/usecases/SetDebuggerEntityUseCase.test.tsx
