+++ src/core/contextual-ui/usecases/SetDebuggerEntityUseCase.tsx
